CTCLoss - reshaping tensors internally to avoid costly exchanges
Summary:
Rearrange shapes of data,gradient and associated exchange and alpha/beta working tensors to have time as the 2nd innermost dimension.  This avoids costly rearrangements when mapping >1 item from the batch onto the same tile.

Reshaped count to be consistent.

We could better describe the order of these shapes with definitions of some sort, but it's difficult to think how this makes it brainless to swap them, when the order they are listed in arrays matter. Although a nice idea, for another day I think.

Ref T32938

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D40839

+++ lib/popnn/codelets/CTCLoss.cpp
@@ -72,7 +72,7 @@ public:
   // within the whole input.  Only validTime (of the whole input) is to be
   // processed. This may mean it has nothing to do
   Input<unsigned> validTime;
-  InOut<unsigned short> count;
+  InOut<unsigned> count;
   const unsigned short labelOffset;
   const unsigned short timeOffset;
   const unsigned short maxT;
@@ -206,7 +206,7 @@ public:
   // within the whole input.  Only validTime (of the whole input) is to be
   // processed. This may mean it has nothing to do
   Input<unsigned> validTime;
-  InOut<unsigned short> count;
+  InOut<unsigned> count;
   const unsigned short labelOffset;
   const unsigned short timeOffset;
   const unsigned short maxT;
@@ -357,7 +357,7 @@ public:
   // within the whole input.  Only validTime (of the whole input) is to be
   // processed. This may mean it has nothing to do
   Input<unsigned> validTime;
-  InOut<unsigned short> count;
+  InOut<unsigned> count;
   const unsigned short labelOffset;
   const unsigned short timeOffset;
   const unsigned short maxT;
@@ -531,7 +531,7 @@ public:
   // within the whole input.  Only validTime (of the whole input) is to be
   // processed. This may mean it has nothing to do
   Input<unsigned> validTime;
-  InOut<unsigned short> count;
+  InOut<unsigned> count;
   const unsigned short labelOffset;
   const unsigned short timeOffset;
   const unsigned short maxT;
